type match for const

+++ contract-nft/src/mint.rs
@@ -3,7 +3,7 @@ use near_sdk::{log, env};
 use near_sdk::json_types::Base64VecU8;
 
 const MAXOUT_USER_NFTS: i8 = 24;
-const MAXOUT_CONTRACT_NTFS: i16 = 100000;
+const MAXOUT_CONTRACT_NTFS: i32 = 100000;
 
 #[near_bindgen]
 impl Contract {
